+++ src/components/calorie-tracker/PhotoUpload.tsx
@@ -5,11 +5,11 @@ import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Input } from '@/components/ui/input'
 import { Alert, AlertDescription } from '@/components/ui/alert'
-import { Camera, Upload, FileImage } from 'lucide-react'
+import { Camera, FileImage } from 'lucide-react'
 import CameraCapture from './CameraCapture'
 import ImagePreview from './ImagePreview'
 import NutritionEntryForm from './NutritionEntryForm'
-import { SupabaseStorageClient } from '@/lib/supabase-storage'
+import { SupabaseStorageClient } from '@/lib/supabase-storage-client'
 import { ImageProcessor } from '@/lib/image-processing'
 import { FileValidator } from '@/utils/file-validation'
 import { UploadState, NutritionLogInput } from '@/lib/nutrition-types'
