+++ src/lib/inngest/functions/analyze-food.ts
@@ -3,7 +3,7 @@ import { NutritionAnalyzer } from '@/lib/ai/nutrition-analyzer';
 import { createClient } from '@/utils/supabase/server';
 import logger from '@/lib/logger';
 
-export const analyzeFoodImage = inngest.createFunction(
+export const analyzeFoodImage = inngest!.createFunction(
   { id: 'analyze-food-image' },
   { event: 'food/image.uploaded' },
   async ({ event, step }) => {
