+++ src/lib/inngest/functions/error-handler.ts
@@ -3,7 +3,7 @@ import { createClient } from '@/utils/supabase/server';
 import { AIErrorMonitor } from '@/lib/monitoring/ai-errors';
 import logger from '@/lib/logger';
 
-export const handleAnalysisError = inngest.createFunction(
+export const handleAnalysisError = inngest!.createFunction(
   { id: 'handle-analysis-error' },
   { event: 'food/analysis.failed' },
   async ({ event, step }) => {
